chore: export factory from size
